Fully operational! One can now move from town to town and it is just so fabulous. There are two new things that remain to be done:
- if logging onto the new server fails, we need to try logging back onto our
old server (and if that fails, we need to throw up our hands in despair; or the
digital equivalent in any case).

- we need to make sure that Getdown has been configured with and downloaded the
extra resources needed for the new town, if it hasn't we need to do some
jiggery pokery and restart the client.

+++ src/java/com/threerings/bang/station/server/StationManager.java
@@ -108,10 +108,16 @@ public class StationManager extends PlaceManager
         }
 
         protected void persistentAction () throws PersistenceException {
+            BangServer.playrepo.grantTownAccess(
+                _user.playerId, _ticket.getTownId());
             BangServer.itemrepo.insertItem(_ticket);
         }
         protected void rollbackPersistentAction () throws PersistenceException {
-            BangServer.itemrepo.deleteItem(_ticket, "ticket_rollback");
+            String oldTownId = BangCodes.TOWN_IDS[_ticket.getTownIndex()-1];
+            BangServer.playrepo.grantTownAccess(_user.playerId, oldTownId);
+            if (_ticket.getItemId() > 0) {
+                BangServer.itemrepo.deleteItem(_ticket, "ticket_rollback");
+            }
         }
 
         protected void actionCompleted () {
